requirements.txt
altgraph==0.17.5
Deprecated==1.3.1
img2pdf==0.6.3
lxml==6.0.2
natsort==8.4.0
packaging==25.0
pefile==2024.8.26
pikepdf==10.2.0
pillow==12.1.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
pywin32-ctypes==0.2.3
wrapt==2.0.1
